# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DHP Enterprise Service API",
    "description": "Swagger definition for Digital Health Pass Enterprise Service API",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "/api/v1/enterprise",
      "description": "VPC"
    },
    {
      "url": "/",
      "description": "Sandbox"
    }
  ],
  "paths": {
...
    "/organization/register": {
      "post": {
        "tags": [
          "organizations"
        ],
        "summary": "Register organization",
        "description": "Register an organization for enterprise services",
        "operationId": "registerOrganization",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Details of organization to be registered including its ID and configuration",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Organization"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully registered organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Failed registering organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicErrorMessage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/organization": {
      "put": {
        "tags": [
          "organizations"
        ],
        "summary": "Update organization",
        "description": "Update an organization already registered for enterprise services",
        "operationId": "updateOrganization",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Details of organization to be updated including its ID and configuration changes",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Organization"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicMessage"
                }
              }
            }
          },
          "400": {
            "description": "Failed updating organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicErrorMessage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/organization/{orgId}": {
      "get": {
        "tags": [
          "organizations"
        ],
        "summary": "Get organization",
        "description": "Get details of specified organization",
        "operationId": "getOrganization",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orgId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Organization ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidGetRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "organizations"
        ],
        "summary": "Delete organization",
        "description": "Delete specified organization",
        "operationId": "deleteOrganization",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orgId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Organization ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicMessage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidGetRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/job": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Create job",
        "description": "Create an enterprise services job for an organization",
        "operationId": "createJob",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Details of job to be created including organization and job IDs as well as status",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Job"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "400": {
            "description": "Failed creating job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicErrorMessage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "put": {
        "tags": [
          "jobs"
        ],
        "summary": "Update job",
        "description": "Update an enterprise services job already created for an organization",
        "operationId": "updateJob",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Details of job to be updated including organization and job IDs as well as status change",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Job"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicMessage"
                }
              }
            }
          },
          "400": {
            "description": "Failed updating job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicErrorMessage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Organization": {
        "type": "object",
        "required": [
          "orgId",
          "config"
        ],
        "properties": {
          "orgId": {
            "type": "string",
            "description": "Organization ID"
          },
          "config": {
            "type": "object",
            "description": "Organization configuration properties"
          }
        }
      },
      "OrganizationResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "payload": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Organization"
              },
              {
                "type": "object",
                "properties": {
                  "createdAt": {
                    "type": "string"
                  },
                  "updatedAt": {
                    "type": "string"
                  }
                }
              }
            ]
          }
        }
      },
      "Job": {
        "type": "object",
        "required": [
          "orgId",
          "jobId",
          "status"
        ],
        "properties": {
          "orgId": {
            "type": "string",
            "description": "Organization ID"
          },
          "jobId": {
            "type": "string",
            "description": "Job ID"
          },
          "status": {
            "type": "string",
            "description": "Job status"
          }
        }
      },
      "JobResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "payload": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Job"
              },
              {
                "type": "object",
                "properties": {
                  "createdAt": {
                    "type": "string"
                  },
                  "updatedAt": {
                    "type": "string"
                  }
                }
              }
            ]
          }
        }
      },
...
      "BasicErrorMessage": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
      "BasicMessage": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
...
      "InvalidGetRequestError": {
        "description": "Invalid fields in request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "Logged in user is unauthorized to make this request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Resource was not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Unexpected server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      }
    }
  }
}